1. fix a possible panic in case a message with larger than MAX_U16 polynomial is sent 2. add more comments 3. change Nodes::reduce to new_reduced so it can only be applied once

+++ fastcrypto-tbls/src/dkg.rs
@@ -142,10 +142,10 @@ pub struct Output<G: GroupElement, EG: GroupElement> {
 ///
 /// Can be instantiated with G1Curve or G2Curve.
 impl<G, EG> Party<G, EG>
-where
-    G: GroupElement + MultiScalarMul + Serialize + DeserializeOwned,
-    EG: GroupElement + Serialize + DeserializeOwned,
-    EG::ScalarType: FiatShamirChallenge,
+    where
+        G: GroupElement + MultiScalarMul + Serialize + DeserializeOwned,
+        EG: GroupElement + Serialize + DeserializeOwned,
+        EG::ScalarType: FiatShamirChallenge,
 {
     /// 1. Create a new ECIES private key and send the public key to all parties.
     /// 2. After *all* parties have sent their ECIES public keys, create the (same) set of nodes.
@@ -270,7 +270,7 @@ where
             return Err(FastCryptoError::InvalidMessage);
         };
 
-        if self.t != msg.vss_pk.degree() + 1 {
+        if self.t as usize != msg.vss_pk.degree() + 1 {
             warn!(
                 "DKG: Message sanity check failed for id {}, expected degree={}, got {}",
                 msg.sender,
@@ -563,7 +563,7 @@ where
                             )),
                             rng,
                         )
-                        .is_ok()
+                            .is_ok()
                     }
                     None => false,
                 };
